Finished Logic and Functionality

+++ encyclopedia/util.py
@@ -1,4 +1,5 @@
 import re
+import markdown2
 
 from django.core.files.base import ContentFile
 from django.core.files.storage import default_storage
@@ -13,17 +14,21 @@ def list_entries():
                 for filename in filenames if filename.endswith(".md")))
 
 
-def save_entry(title, content):
+def save_entry(title, content, edit=False):
     """
     Saves an encyclopedia entry, given its title and Markdown
     content. If an existing entry with the same title already exists,
     it is replaced.
     """
     filename = f"entries/{title}.md"
-    if default_storage.exists(filename):
-        default_storage.delete(filename)
-    default_storage.save(filename, ContentFile(content))
 
+    if not edit and default_storage.exists(filename):
+        raise FileExistsError("Entry Already Exists")
+    
+    if edit and default_storage.exists(filename):
+        default_storage.delete(filename)
+
+    default_storage.save(filename, ContentFile(content))
 
 def get_entry(title):
     """
@@ -31,7 +36,20 @@ def get_entry(title):
     entry exists, the function returns None.
     """
     try:
-        f = default_storage.open(f"entries/{title}.md")
-        return f.read().decode("utf-8")
+        with open(f"entries/{title}.md", 'r') as file:
+            return file.read()  
+    except FileNotFoundError:
+        return None
+
+
+def get_raw_entry(title):
+    """
+    Retrieves the raw Markdown content of an encyclopedia entry by its title.
+    If no such entry exists, the function returns None.
+    """
+
+    try:
+        with default_storage.open(f"entries/{title}.md") as file:
+            return file.read()
     except FileNotFoundError:
         return None
